Project_5/requirements.txt.txt
PyQt5==5.15.9
Pillow==10.0.0